improvement logging

+++ src/DirectorySync.Application/UseCases/SynchronizeGroupsUseCase.cs
@@ -1,3 +1,6 @@
+using System.Collections.ObjectModel;
+using DirectorySync.Application.Extensions;
+using DirectorySync.Application.Measuring;
 using DirectorySync.Application.Models.Core;
 using DirectorySync.Application.Models.Enums;
 using DirectorySync.Application.Models.ValueObjects;
@@ -26,6 +29,7 @@ public class SynchronizeGroupsUseCase : ISynchronizeGroupsUseCase
     private readonly IUserDeleter _userDeleter;
     private readonly IGroupUpdater _groupUpdater;
     private readonly ISyncSettingsOptions _syncSettingsOptions;
+    private readonly CodeTimer _codeTimer;
     private readonly ILogger<SynchronizeGroupsUseCase> _logger;
 
     public SynchronizeGroupsUseCase(IGroupDatabase groupDatabase,
@@ -38,6 +42,7 @@ public class SynchronizeGroupsUseCase : ISynchronizeGroupsUseCase
         IUserDeleter userDeleter,
         IGroupUpdater groupUpdater,
         ISyncSettingsOptions syncSettingsOptions,
+        CodeTimer codeTimer,
         ILogger<SynchronizeGroupsUseCase> logger)
     {
         _groupDatabase = groupDatabase;
@@ -50,17 +55,21 @@ public class SynchronizeGroupsUseCase : ISynchronizeGroupsUseCase
         _userDeleter = userDeleter;
         _groupUpdater = groupUpdater;
         _syncSettingsOptions = syncSettingsOptions;
+        _codeTimer = codeTimer;
         _logger = logger;
     }
 
     public async Task ExecuteAsync(IEnumerable<DirectoryGuid> trackingGroupGuids,
         CancellationToken cancellationToken = default)
     {
+        _logger.LogInformation(ApplicationEvent.StartUserScanning, "Start synchronization for groups: {group}", trackingGroupGuids);
+        
         var memberMap = new Dictionary<DirectoryGuid, MemberModel>();
 
         foreach (var groupId in trackingGroupGuids)
         {
-            await ProcessGroupChanges(groupId, memberMap, cancellationToken);
+            using var withGroup = _logger.EnrichWithGroup(groupId);
+            ProcessGroupChanges(groupId, memberMap, cancellationToken);
         }
 
         var toCreate = memberMap.Values
@@ -72,42 +81,79 @@ public class SynchronizeGroupsUseCase : ISynchronizeGroupsUseCase
         var toDelete = memberMap.Values
             .Where(m => m.Operation == ChangeOperation.Delete)
             .ToList();
+
+        ReadOnlyCollection<MemberModel> created = ReadOnlyCollection<MemberModel>.Empty;
+        ReadOnlyCollection<MemberModel> updated = ReadOnlyCollection<MemberModel>.Empty;
+        ReadOnlyCollection<MemberModel> deleted = ReadOnlyCollection<MemberModel>.Empty;
         
-        var created = await _userCreator.CreateManyAsync(toCreate, cancellationToken);
-        var updated = await _userUpdater.UpdateManyAsync(toUpdate, cancellationToken);
-        var deleted = await _userDeleter.DeleteManyAsync(toDelete, cancellationToken);
+        if (toCreate.Count != 0)
+        {
+            created = await _userCreator.CreateManyAsync(toCreate, cancellationToken);
+        }
+
+        if (toUpdate.Count != 0)
+        {
+            updated = await _userUpdater.UpdateManyAsync(toUpdate, cancellationToken);
+        }
+
+        if (toDelete.Count != 0)
+        {
+            deleted = await _userDeleter.DeleteManyAsync(toDelete, cancellationToken);
+        }
         
         _groupUpdater.UpdateGroupsWithMembers(created.Concat(updated).Concat(deleted));
+        
+        _logger.LogInformation(ApplicationEvent.CompleteUserScanning, "Complete groups synchronization");
     }
     
-    private async Task ProcessGroupChanges(DirectoryGuid groupId,
+    private void ProcessGroupChanges(DirectoryGuid groupId,
         Dictionary<DirectoryGuid, MemberModel> memberMap,
         CancellationToken cancellationToken)
     {
+        var getGroupTimer = _codeTimer.Start("Get Reference Group");
         var referenceGroup = _groupPort.GetByGuidAsync(groupId);
+        getGroupTimer.Stop();
         if (referenceGroup is null)
         {
+            _logger.LogWarning("Reference group not found: {Group:l}", groupId);
             return;
         }
+        _logger.LogDebug("Reference group found: {Group:l}", referenceGroup);
 
         var cachedGroup = _groupDatabase.FindById(groupId);
 
         if (cachedGroup is null)
         {
+            _logger.LogDebug("Reference group {Group} is not cached and now it will", groupId);
             cachedGroup = GroupModel.Create(referenceGroup.Id, []);
             _groupDatabase.Insert(cachedGroup);
         }
 
         if (cachedGroup.MembersHash == referenceGroup.MembersHash)
         {
+            _logger.LogDebug("Group {Group} has no changes", groupId);
             return;
         }
 
         var removedIds = cachedGroup.MemberIds.Except(referenceGroup.MemberIds).ToArray();
-        var addedIds = referenceGroup.MemberIds.Except(cachedGroup.MemberIds).ToArray();
+        if (removedIds.Length == 0)
+        {
+            _logger.LogDebug("Removed users not found...");
+        }
+        else
+        {
+            HandleRemovedMembers(groupId, removedIds, memberMap);   
+        }
         
-        HandleRemovedMembers(groupId, removedIds, memberMap);
-        await HandleAddedMembers(groupId, addedIds, memberMap, cancellationToken);
+        var addedIds = referenceGroup.MemberIds.Except(cachedGroup.MemberIds).ToArray();
+        if (addedIds.Length == 0)
+        {
+            _logger.LogDebug("Added users not found...");
+        }
+        else
+        {
+            HandleAddedMembers(groupId, addedIds, memberMap, cancellationToken);
+        }
 
         SetMembersGroupMapping(memberMap.Values);
     }
@@ -133,7 +179,7 @@ public class SynchronizeGroupsUseCase : ISynchronizeGroupsUseCase
         }
     }
     
-    private async Task HandleAddedMembers(DirectoryGuid groupId,
+    private void HandleAddedMembers(DirectoryGuid groupId,
         IEnumerable<DirectoryGuid> addedIds,
         Dictionary<DirectoryGuid, MemberModel> memberMap,
         CancellationToken cancellationToken)
